equity_tracker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "equity-tracker"
version = "0.1.0"
description = "Personal equity compensation and investment tracker — UK, privacy-first"
requires-python = ">=3.12"
dependencies = [
    "SQLAlchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "yfinance>=0.2.40",
    "gspread>=6.1.0",
    "argon2-cffi>=23.1.0",
    # Web API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    # SQLCipher (encrypted SQLite) — installed separately due to native build requirements.
    # Run:  pip install sqlcipher3-binary   (pre-built Windows wheel)
    # Fallback for development without encryption: standard sqlite3 is used automatically.
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -v"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]
ignore = ["N818"]  # Exception naming convention — not applicable to domain errors

[tool.mypy]
python_version = "3.12"
strict = true
